requirements.txt
Flask
pandas
gunicorn
pyodbc
python-dotenv